requirements.txt
feedparser>=6.0.10
beautifulsoup4>=4.12.0
requests>=2.31.0
anthropic>=0.18.0
google-generativeai>=0.4.0
jinja2>=3.1.0
weasyprint>=60.0
PyYAML>=6.0
python-dotenv>=1.0.0
